+++ sys/dev/ebpf/ebpf_map.h
@@ -21,8 +21,8 @@
 
 struct ebpf_map;
 
-typedef int ebpf_map_init_t(struct ebpf_map *self, uint16_t key_size,
-			    uint16_t value_size, uint16_t max_entries,
+typedef int ebpf_map_init_t(struct ebpf_map *self, uint32_t key_size,
+			    uint32_t value_size, uint32_t max_entries,
 			    uint32_t flags);
 typedef void *ebpf_map_lookup_elem_t(struct ebpf_map *self, void *key,
 				     uint64_t flags);
@@ -57,8 +57,8 @@ struct ebpf_map {
 };
 
 void ebpf_register_map_type(uint16_t id, struct ebpf_map_ops *ops);
-int ebpf_map_init(struct ebpf_map *mapp, uint16_t type, uint16_t key_size,
-		  uint16_t value_size, uint32_t max_entries,
+int ebpf_map_init(struct ebpf_map *mapp, uint16_t type, uint32_t key_size,
+		  uint32_t value_size, uint32_t max_entries,
 		  uint32_t map_flags);
 void *ebpf_map_lookup_elem(struct ebpf_map *self, void *key, uint64_t flags);
 int ebpf_map_update_elem(struct ebpf_map *self, void *key, void *value,
