+++ sys/dev/ebpf/ebpf_map_bad.c
@@ -17,8 +17,8 @@
 #include "ebpf_map.h"
 
 static int
-bad_map_init(struct ebpf_map *self, uint16_t key_size, uint16_t value_size,
-	     uint16_t max_entries, uint32_t flags)
+bad_map_init(struct ebpf_map *self, uint32_t key_size, uint32_t value_size,
+	     uint32_t max_entries, uint32_t flags)
 {
 	return EINVAL;
 }
